funcionamiente reviews

+++ demo/src/main/java/com/example/demo/Repository/ReviewRepository.java
@@ -6,8 +6,6 @@ import com.example.demo.Model.Review;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-
-    // Opcional: por si quieres obtener reviews de un producto concreto
-    List<Review> findByProductId(Integer productId);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    List<Review> findByProductId(int productId);
 }
